Split gateway_tests and authority_aggregator_tests
